+++ src/models/user.model.ts
@@ -16,9 +16,9 @@ export default class UserModel {
     #createTableIfNotExists () {
         this.client.query(`            
             CREATE TABLE IF NOT EXISTS users (
-                ID SERIAL PRIMARY KEY,
+                ID BIGSERIAL PRIMARY KEY,
                 FULLNAME VARCHAR(64) DEFAULT 'student',
-                TELEGRAM_USER_ID INT NOT NULL UNIQUE,
+                TELEGRAM_USER_ID BIGINT NOT NULL UNIQUE,
                 CONTACT VARCHAR(32) NOT NULL UNIQUE,
                 ROLE user_role DEFAULT 'student', 
                 SIGNED_TIME TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP
